Move view file to view/settings + code split
